+++ from_pouet.net/with_source_032/ia32_com/by_reko.c
@@ -24,7 +24,7 @@ void fn0C00_0100(byte Eq_4::* si, byte Eq_25::* di, struct Eq_4 * ds)
 			ah_14 = SLICE(ax_15 - 0x01, byte, 8);
 			di = di_21 + 1;
 		} while (di_21 != &Eq_25::b0001);
-		uint16 ax_29 = (uint16) (uint8) __inb(0x60);
+		uint16 ax_29 = (uint16) __inb(0x60);
 		ah_14 = SLICE(ax_29 - 0x01, byte, 8);
 	} while (ax_29 != 0x01);
 }
