+++ from_pouet.net/with_source_034/ia32_com/by_reko.c
@@ -11,7 +11,7 @@ void fn0C00_0100(struct Eq_2 * ds_bx, word16 cx, byte Eq_10::* di)
 	struct Eq_10 * ds_37 = ds_bx->ptr0002;
 	while (true)
 	{
-		uint32 dx_ax_13 = (uint32) (uint16) di;
+		uint32 dx_ax_13 = (uint32) di;
 		ds_37->*di = ds_37->*di ^ (byte) (((uint16) (dx_ax_13 /u 0x0140) + cx & (uint16) (dx_ax_13 % 0x0140) + cx) >> (byte) cx);
 		++di;
 		if (di == &Eq_10::b0000)
